+++ VComponents/Components/Buttons/Secondary Button/VSecondaryButtonModel.swift
@@ -56,10 +56,7 @@ public struct VSecondaryButtonModel {
         public var iconTitleSpacing: CGFloat = 8
         
         /// Hit box. Defaults to `0` horizontally and `0` vertically.
-        public var hitBox: HitBox = .init(
-            horizontal: 0,
-            vertical: 0
-        )
+        public var hitBox: HitBox = .zero
         
         // MARK: Initializers
         /// Initializes sub-model with default values.