+++ VComponents/Components/Lists/Base List/VListLayoutType.swift
@@ -1,5 +1,5 @@
 //
-//  VGenericListContentLayout.swift
+//  VListLayoutType.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 1/11/21.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-// MARK: - V Base List Layout Type
+// MARK: - V List Layout Type
 /// Enum that describes layout, such as `fixed` or `flexible`.
 ///
 /// There are three posible layouts:
@@ -20,7 +20,7 @@ import Foundation
 ///
 /// 3. `Constrained`.
 /// `.frame()` modifier can be applied to view. Content would be limitd in vertical space. Scrolling may be enabled inside component.
-public enum VBaseListLayoutType: Int, CaseIterable {
+public enum VListLayoutType: Int, CaseIterable {
     // MARK: Cases
     /// Fixed layout.
     ///